requirements.txt
asgiref==3.8.1
Django==5.2
django-unfold==0.54.0
pillow==11.1.0
sqlparse==0.5.3
